Reorganize segmentation parsers
By putting them all in separate modules, we communicate that a new
parser can and should be started in a new module, independently from the
others.
