+++ lib/segmentation/merizo.py
@@ -1,0 +1,6 @@
+from . import SegmentationParser
+
+
+class Parser(SegmentationParser):
+    def __init__(self, path):
+        super().__init__(path)
